# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: ADA ASM API
  version: 0.1.0
  description: |
    REST API for ADA ASM. This file is the source of truth for the API
    contract and is regenerated from the FastAPI app's `/openapi.json`
    on every PR that changes `backend/app/api/**`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/projects:
    get:
      tags: [projects]
      summary: List projects (catalogue), aggregates hydrated server-side
      description: |
        Default behaviour excludes `Archived` rows. Pass `include_archived=true`
        to lift the exclusion. `status[]` is repeatable and, when provided,
        replaces the default filter entirely.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Case-insensitive over `code`, `name`, `customer.name`.
        - in: query
          name: status
          schema:
            type: array
            items: { $ref: "#/components/schemas/ProjectStatus" }
          style: form
          explode: true
        - in: query
          name: include_archived
          schema: { type: boolean, default: false }
        - in: query
          name: customer_id
          schema: { type: array, items: { type: string, format: uuid } }
          style: form
          explode: true
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: PaginatedProjects — each item carries aggregates + customer hydrated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProjects" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreate" }
      responses:
        "201":
          description: ProjectResponse (hydrated)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "409":
          description: PROJECT_CODE_ALREADY_REGISTERED
        "422":
          description: Validation errors.
  /api/v1/projects/{project_id}:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [projects]
      summary: Project detail (aggregates + children + customer hydrated)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ProjectResponse
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "404":
          description: PROJECT_NOT_FOUND
    patch:
      tags: [projects]
      summary: Partial update — code is mutable; transition to Delivered auto-fills fecha_entrega_real
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: ProjectResponse
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "404":
          description: PROJECT_NOT_FOUND
        "409":
          description: PROJECT_CODE_ALREADY_REGISTERED
    delete:
      tags: [projects]
      summary: Soft-delete — sets status='Archived'
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Archived (idempotent)
        "404":
          description: PROJECT_NOT_FOUND
  /api/v1/projects/{project_id}/children:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [projects]
      summary: Add a project→(module|component) edge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddProjectChildRequest" }
      responses:
        "201":
          description: ProjectChildResponse (hydrated)
        "404":
          description: PROJECT_NOT_FOUND or INVALID_CHILD_REFERENCE
        "409":
          description: CHILD_ALREADY_PRESENT
        "422":
          description: XOR violation or validation errors.
  /api/v1/projects/{project_id}/children/{child_id}:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: child_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [projects]
      summary: Patch quantity / notes / sort_order on an existing edge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectChildRequest" }
      responses:
        "200":
          description: ProjectChildResponse (re-hydrated)
        "404":
          description: PROJECT_NOT_FOUND or child missing
    delete:
      tags: [projects]
      summary: Remove an edge (referenced entity is NOT touched)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Removed (idempotent)
  /api/v1/projects/{project_id}/price-history:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
      - in: query
        name: period
        schema: { type: string, enum: [week, month, year], default: year }
    get:
      tags: [projects]
      summary: Aggregated price time-series (server-computed)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ProjectPriceHistoryResponse
        "404":
          description: PROJECT_NOT_FOUND
  /api/v1/projects/{project_id}/stock-events:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [projects]
      summary: Stock events tied to this project (consumption + delivered)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Paginated StockEvent rows ordered by `occurred_at DESC`.
        "404":
          description: PROJECT_NOT_FOUND
  /api/v1/projects/{project_id}/interest-links:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [projects]
      summary: Create an interest link for a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInterestLinkCreate" }
      responses:
        "201":
          description: ProjectInterestLink
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectInterestLink" }
        "404":
          description: PROJECT_NOT_FOUND
        "422":
          description: Validation errors.
  /api/v1/projects/{project_id}/interest-links/{link_id}:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: link_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [projects]
      summary: Patch an interest link (name / url / sort_order)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInterestLinkUpdate" }
      responses:
        "200":
          description: ProjectInterestLink
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectInterestLink" }
        "404":
          description: PROJECT_NOT_FOUND or link missing
    delete:
      tags: [projects]
      summary: Remove an interest link
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Removed (idempotent on missing)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectStatus:
      type: string
      enum: [Presupuestado, Esperando, En proceso, Completado, Archivado]
      description: |
        Lifecycle status — DB-enforced (CHECK) and shared with the FE labels verbatim.
        Soft-delete transitions to `Archivado`. The PATCH endpoint auto-fills
        `fecha_entrega_real` with today's date when the status changes to
        `Completado` and the body does not provide an explicit value.
...
    ProjectResponse:
      type: object
      description: Detail bundle — ProjectSummary + hydrated children + interest links.
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          required: [children, interest_links]
          properties:
            children:
              type: array
              items: { $ref: "#/components/schemas/ProjectChild" }
            interest_links:
              type: array
              items: { $ref: "#/components/schemas/ProjectInterestLink" }
...
    ProjectInterestLinkCreate:
      type: object
      required: [name, url]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        url: { type: string, minLength: 1, maxLength: 2000 }
        sort_order: { type: integer }
    ProjectInterestLinkUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        url: { type: string, minLength: 1, maxLength: 2000 }
        sort_order: { type: integer }
    PaginatedProjects:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ProjectSummary" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
    ProjectCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customer_id: { type: string, format: uuid, nullable: true }
        icon: { type: string, nullable: true, maxLength: 8 }
        color: { type: string, nullable: true, maxLength: 7 }
        tags:
          type: array
          items: { type: string }
        version: { type: string, nullable: true, maxLength: 40 }
        fecha_inicio: { type: string, format: date, nullable: true }
        fecha_entrega_estimada: { type: string, format: date, nullable: true }
        fecha_entrega_real: { type: string, format: date, nullable: true }
        notas: { type: string, nullable: true }
    ProjectUpdate:
      type: object
      description: All fields optional; `code` IS mutable.
      properties:
        code: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProjectStatus" }
        customer_id: { type: string, format: uuid, nullable: true }
        icon: { type: string, nullable: true, maxLength: 8 }
        color: { type: string, nullable: true, maxLength: 7 }
        tags:
          type: array
          items: { type: string }
        version: { type: string, nullable: true, maxLength: 40 }
        fecha_inicio: { type: string, format: date, nullable: true }
        fecha_entrega_estimada: { type: string, format: date, nullable: true }
        fecha_entrega_real: { type: string, format: date, nullable: true }
        notas: { type: string, nullable: true }
    AddProjectChildRequest:
      type: object
      required: [quantity]
      properties:
        child_module_id: { type: string, format: uuid }
        child_component_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1 }
        notes: { type: string, nullable: true }
        sort_order: { type: integer }
    UpdateProjectChildRequest:
      type: object
      properties:
        quantity: { type: integer, minimum: 1 }
        notes: { type: string, nullable: true }
        sort_order: { type: integer }
...
  responses:
    Unauthorized:
      description: |
        Authentication failed. `code` is one of:
        `INVALID_CREDENTIALS`, `UNAUTHENTICATED`, `ACCESS_TOKEN_EXPIRED`,
        `ACCESS_TOKEN_WRONG_TYPE`, `ACCESS_TOKEN_MALFORMED`,
        `REFRESH_TOKEN_EXPIRED`, `REFRESH_TOKEN_REVOKED`,
        `REFRESH_TOKEN_MALFORMED`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...